[#42] - cast cache lifetime to int

+++ library/Core/config.php
@@ -33,7 +33,9 @@ return [
             'host'     => envValue('DATA_API_REDIS_HOST', '127.0.0.1'),
             'port'     => envValue('DATA_API_REDIS_PORT', 6379),
             'index'    => envValue('DATA_API_REDIS_WEIGHT', 0),
-            'lifetime' => envValue('CACHE_LIFETIME', 86400),
+            // Cast: env values arrive as strings, and AbstractAdapter::getTtl()
+            // is typed to return int.
+            'lifetime' => (int) envValue('CACHE_LIFETIME', 86400),
             'prefix'   => 'data-',
         ],
     ],
@@ -48,7 +50,9 @@ return [
                 'host'     => envValue('DATA_API_REDIS_HOST', '127.0.0.1'),
                 'port'     => envValue('DATA_API_REDIS_PORT', 6379),
                 'index'    => envValue('DATA_API_REDIS_WEIGHT', 0),
-                'lifetime' => envValue('CACHE_LIFETIME', 86400),
+                // Cast: env values arrive as strings, and AbstractAdapter::getTtl()
+                // is typed to return int.
+                'lifetime' => (int) envValue('CACHE_LIFETIME', 86400),
                 'prefix'   => 'metadata-',
             ],
         ],
